Fixed UI calibration
- Projectiles shooting still wrong, they're turning clockwisely with
counter-clockwise angle :(
- And radial input was fixed

+++ The_Erder_Dedicated_New/src/ru/alastar/game/Entity.java
@@ -715,7 +715,7 @@ public class Entity extends Transform implements IUpdate, IPhysic
         return pData;
     }
 
-    public void act(double angle)
+    public void act(float angle)
     {
         if(warMode)
         {
