+++ The_Erder_Dedicated_New/src/ru/alastar/main/net/requests/ActInput.java
@@ -2,5 +2,5 @@ package ru.alastar.main.net.requests;
 
 public class ActInput
 {
-  public double angle;
+  public float angle;
 }
