+++ the-erder-core/src/com/alastar/game/gui/GUIChatWindow.java
@@ -25,16 +25,16 @@ public class GUIChatWindow implements GUIElement
         this.window = new Window(this.name,
                 GameManager.getSkin(GameManager.selectedSkin), "window");
         this.window.setName(name);
-        this.window.setPosition(0 / Vars.getInt("balancedScreenWidth"),
-                0 / Vars.getInt("balancedScreenHeight"));
-        this.window.setHeight(0 / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(0 / Vars.getInt("balancedScreenWidth"));
+        this.window.setPosition(0 / (float)Vars.getDouble("balancedScreenWidth"),
+                0 / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setHeight(0 / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(0 / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(0);
         window.defaults().padRight(0);
         window.defaults().padTop(0);
         window.defaults().padBottom(0);
-        window.defaults().minWidth(100 / Vars.getInt("balancedScreenWidth"));
-        window.defaults().minHeight(30 / Vars.getInt("balancedScreenHeight"));
+        window.defaults().minWidth(100 / (float)Vars.getDouble("balancedScreenWidth"));
+        window.defaults().minHeight(30 / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     public GUIChatWindow(String n, Window w, Vector2 vector2, Vector2 vector22,
@@ -43,18 +43,18 @@ public class GUIChatWindow implements GUIElement
         this.name = n;
         this.window = w;
         this.window.setName(n);
-        this.window.setPosition(vector2.x / Vars.getInt("balancedScreenWidth"),
-                vector2.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setHeight(vector22.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(vector22.x / Vars.getInt("balancedScreenWidth"));
+        this.window.setPosition(vector2.x / (float)Vars.getDouble("balancedScreenWidth"),
+                vector2.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setHeight(vector22.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(vector22.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(i);
         window.defaults().padRight(j);
         window.defaults().padTop(k);
         window.defaults().padBottom(l);
         window.defaults().prefWidth(
-                vector22.x / Vars.getInt("balancedScreenWidth"));
+                vector22.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().prefHeight(
-                vector22.y / Vars.getInt("balancedScreenHeight"));
+                vector22.y / (float)Vars.getDouble("balancedScreenHeight"));
 
         table = new Table();
         table.defaults().padLeft(i);
@@ -162,7 +162,7 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setHeight(float val)
     {
-        this.window.setHeight(val / Vars.getInt("balancedScreenHeight"));
+        this.window.setHeight(val / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -174,7 +174,7 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setWidth(float val)
     {
-        this.window.setWidth(val / Vars.getInt("balancedScreenWidth"));
+        this.window.setWidth(val / (float)Vars.getDouble("balancedScreenWidth"));
 
     }
 
@@ -187,8 +187,8 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setPosition(Vector2 val)
     {
-        window.setX(val.x / Vars.getInt("balancedScreenWidth"));
-        window.setY(val.y / Vars.getInt("balancedScreenHeight"));
+        window.setX(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        window.setY(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -200,9 +200,9 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setPadTB(Vector2 val)
     {
-        window.defaults().padTop(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().padTop(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .padBottom(val.y / Vars.getInt("balancedScreenHeight"));
+                .padBottom(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -214,8 +214,8 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setPadRL(Vector2 val)
     {
-        window.defaults().padRight(val.x / Vars.getInt("balancedScreenWidth"));
-        window.defaults().padLeft(val.y / Vars.getInt("balancedScreenHeight"));
+        window.defaults().padRight(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        window.defaults().padLeft(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -227,9 +227,9 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setMinHW(Vector2 val)
     {
-        window.defaults().minWidth(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().minWidth(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .minHeight(val.y / Vars.getInt("balancedScreenHeight"));
+                .minHeight(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -241,9 +241,9 @@ public class GUIChatWindow implements GUIElement
     @Override
     public void setMaxHW(Vector2 val)
     {
-        window.defaults().maxWidth(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().maxWidth(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .maxHeight(val.y / Vars.getInt("balancedScreenHeight"));
+                .maxHeight(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
