+++ the-erder-core/src/com/alastar/game/gui/GUIEquipSlot.java
@@ -111,7 +111,7 @@ public class GUIEquipSlot implements GUIElement
     @Override
     public void setHeight(float val)
     {
-        this.label.setHeight(val / Vars.getInt("balancedScreenHeight"));
+        this.label.setHeight(val / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -123,7 +123,7 @@ public class GUIEquipSlot implements GUIElement
     @Override
     public void setWidth(float val)
     {
-        this.label.setWidth(val / Vars.getInt("balancedScreenWidth"));
+        this.label.setWidth(val / (float)Vars.getDouble("balancedScreenWidth"));
 
     }
 
@@ -136,8 +136,8 @@ public class GUIEquipSlot implements GUIElement
     @Override
     public void setPosition(Vector2 val)
     {
-        label.setX(val.x / Vars.getInt("balancedScreenWidth"));
-        label.setY(val.y / Vars.getInt("balancedScreenHeight"));
+        label.setX(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        label.setY(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
