+++ the-erder-core/src/com/alastar/game/gui/GUIList.java
@@ -24,16 +24,16 @@ public class GUIList implements GUIElement
         this.window = new Window(this.name,
                 GameManager.getSkin(GameManager.selectedSkin), "window");
         this.window.setName(name);
-        this.window.setPosition(0 / Vars.getInt("balancedScreenWidth"),
-                0 / Vars.getInt("balancedScreenHeight"));
-        this.window.setHeight(0 / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(0 / Vars.getInt("balancedScreenWidth"));
+        this.window.setPosition(0 / (float)Vars.getDouble("balancedScreenWidth"),
+                0 / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setHeight(0 / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(0 / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(0);
         window.defaults().padRight(0);
         window.defaults().padTop(0);
         window.defaults().padBottom(0);
-        window.defaults().minWidth(100 / Vars.getInt("balancedScreenWidth"));
-        window.defaults().minHeight(30 / Vars.getInt("balancedScreenHeight"));
+        window.defaults().minWidth(100 / (float)Vars.getDouble("balancedScreenWidth"));
+        window.defaults().minHeight(30 / (float)Vars.getDouble("balancedScreenHeight"));
 
         table = new Table();
         table.defaults().padLeft(0);
@@ -65,10 +65,10 @@ public class GUIList implements GUIElement
         this.name = n;
         this.window = w;
         this.window.setName(n);
-        this.window.setPosition(vector2.x / Vars.getInt("balancedScreenWidth"),
-                vector2.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setHeight(vector22.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(vector22.x / Vars.getInt("balancedScreenWidth"));
+        this.window.setPosition(vector2.x / (float)Vars.getDouble("balancedScreenWidth"),
+                vector2.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setHeight(vector22.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(vector22.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(i);
         window.defaults().padRight(j);
         window.defaults().padTop(k);
@@ -160,7 +160,7 @@ public class GUIList implements GUIElement
     @Override
     public void setHeight(float val)
     {
-        this.window.setHeight(val / Vars.getInt("balancedScreenHeight"));
+        this.window.setHeight(val / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -172,7 +172,7 @@ public class GUIList implements GUIElement
     @Override
     public void setWidth(float val)
     {
-        this.window.setWidth(val / Vars.getInt("balancedScreenWidth"));
+        this.window.setWidth(val / (float)Vars.getDouble("balancedScreenWidth"));
 
     }
 
@@ -185,8 +185,8 @@ public class GUIList implements GUIElement
     @Override
     public void setPosition(Vector2 val)
     {
-        window.setX(val.x / Vars.getInt("balancedScreenWidth"));
-        window.setY(val.y / Vars.getInt("balancedScreenHeight"));
+        window.setX(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        window.setY(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -198,9 +198,9 @@ public class GUIList implements GUIElement
     @Override
     public void setPadTB(Vector2 val)
     {
-        window.defaults().padTop(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().padTop(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .padBottom(val.y / Vars.getInt("balancedScreenHeight"));
+                .padBottom(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -212,8 +212,8 @@ public class GUIList implements GUIElement
     @Override
     public void setPadRL(Vector2 val)
     {
-        window.defaults().padRight(val.x / Vars.getInt("balancedScreenWidth"));
-        window.defaults().padLeft(val.y / Vars.getInt("balancedScreenHeight"));
+        window.defaults().padRight(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        window.defaults().padLeft(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -225,9 +225,9 @@ public class GUIList implements GUIElement
     @Override
     public void setMinHW(Vector2 val)
     {
-        window.defaults().minWidth(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().minWidth(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .minHeight(val.y / Vars.getInt("balancedScreenHeight"));
+                .minHeight(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -239,9 +239,9 @@ public class GUIList implements GUIElement
     @Override
     public void setMaxHW(Vector2 val)
     {
-        window.defaults().maxWidth(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().maxWidth(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .maxHeight(val.y / Vars.getInt("balancedScreenHeight"));
+                .maxHeight(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
