+++ the-erder-core/src/com/alastar/game/gui/GUIWindow.java
@@ -22,16 +22,16 @@ public class GUIWindow implements GUIElement
         this.window = new Window(this.name,
                 GameManager.getSkin(GameManager.selectedSkin), "window");
         this.window.setName(name);
-        this.window.setPosition(0 / Vars.getInt("balancedScreenWidth"),
-                0 / Vars.getInt("balancedScreenHeight"));
-        this.window.setHeight(0 / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(0 / Vars.getInt("balancedScreenWidth"));
+        this.window.setPosition(0 / (float)Vars.getDouble("balancedScreenWidth"),
+                0 / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setHeight(0 / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(0 / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(0);
         window.defaults().padRight(0);
         window.defaults().padTop(0);
         window.defaults().padBottom(0);
-        window.defaults().minWidth(100 / Vars.getInt("balancedScreenWidth"));
-        window.defaults().minHeight(30 / Vars.getInt("balancedScreenHeight"));
+        window.defaults().minWidth(100 / (float)Vars.getDouble("balancedScreenWidth"));
+        window.defaults().minHeight(30 / (float)Vars.getDouble("balancedScreenHeight"));
 
         mainTable = new Table();
         mainTable.setFillParent(true);
@@ -45,10 +45,10 @@ public class GUIWindow implements GUIElement
         this.name = n;
         this.window = w;
         this.window.setName(n);
-        this.window.setPosition(vector2.x / Vars.getInt("balancedScreenWidth"),
-                vector2.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setHeight(vector22.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(vector22.x / Vars.getInt("balancedScreenWidth"));
+        this.window.setPosition(vector2.x / (float)Vars.getDouble("balancedScreenWidth"),
+                vector2.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setHeight(vector22.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(vector22.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(i);
         window.defaults().padRight(j);
         window.defaults().padTop(k);
@@ -64,8 +64,8 @@ public class GUIWindow implements GUIElement
         this.name = n;
         this.window = w;
         this.window.setName(n);
-        this.window.setHeight(vector22.y / Vars.getInt("balancedScreenHeight"));
-        this.window.setWidth(vector22.x / Vars.getInt("balancedScreenWidth"));
+        this.window.setHeight(vector22.y / (float)Vars.getDouble("balancedScreenHeight"));
+        this.window.setWidth(vector22.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults().padLeft(i);
         window.defaults().padRight(j);
         window.defaults().padTop(k);
@@ -161,7 +161,7 @@ public class GUIWindow implements GUIElement
     @Override
     public void setHeight(float val)
     {
-        this.window.setHeight(val / Vars.getInt("balancedScreenHeight"));
+        this.window.setHeight(val / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -173,7 +173,7 @@ public class GUIWindow implements GUIElement
     @Override
     public void setWidth(float val)
     {
-        this.window.setWidth(val / Vars.getInt("balancedScreenWidth"));
+        this.window.setWidth(val / (float)Vars.getDouble("balancedScreenWidth"));
 
     }
 
@@ -186,8 +186,8 @@ public class GUIWindow implements GUIElement
     @Override
     public void setPosition(Vector2 val)
     {
-        window.setX(val.x / Vars.getInt("balancedScreenWidth"));
-        window.setY(val.y / Vars.getInt("balancedScreenHeight"));
+        window.setX(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        window.setY(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -199,9 +199,9 @@ public class GUIWindow implements GUIElement
     @Override
     public void setPadTB(Vector2 val)
     {
-        window.defaults().padTop(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().padTop(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .padBottom(val.y / Vars.getInt("balancedScreenHeight"));
+                .padBottom(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -213,8 +213,8 @@ public class GUIWindow implements GUIElement
     @Override
     public void setPadRL(Vector2 val)
     {
-        window.defaults().padRight(val.x / Vars.getInt("balancedScreenWidth"));
-        window.defaults().padLeft(val.y / Vars.getInt("balancedScreenHeight"));
+        window.defaults().padRight(val.x / (float)Vars.getDouble("balancedScreenWidth"));
+        window.defaults().padLeft(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -226,9 +226,9 @@ public class GUIWindow implements GUIElement
     @Override
     public void setMinHW(Vector2 val)
     {
-        window.defaults().minWidth(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().minWidth(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .minHeight(val.y / Vars.getInt("balancedScreenHeight"));
+                .minHeight(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
@@ -240,9 +240,9 @@ public class GUIWindow implements GUIElement
     @Override
     public void setMaxHW(Vector2 val)
     {
-        window.defaults().maxWidth(val.x / Vars.getInt("balancedScreenWidth"));
+        window.defaults().maxWidth(val.x / (float)Vars.getDouble("balancedScreenWidth"));
         window.defaults()
-                .maxHeight(val.y / Vars.getInt("balancedScreenHeight"));
+                .maxHeight(val.y / (float)Vars.getDouble("balancedScreenHeight"));
     }
 
     @Override
